fix: detail 페이지 data 가져오는 부분 useQuery로 전환
