+++ src/components/Detail/DetailContent/index.tsx
@@ -1,40 +1,19 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import Button from 'components/Button'
 import { TwoButtonModal } from 'components/Modal'
 import useModal from 'hooks/useModal'
 import { LikeIcon, LikeOnIcon, LocationIcon, PeopleIcon, TimeIcon } from 'assets/svgs'
-import { authInstance } from 'apis/client'
-import { useNavigate } from 'react-router-dom'
+import { datailData } from 'types'
+
 import styles from './detailContent.module.scss'
 
 interface Props {
-  id?: string
+  data?: datailData
 }
 
-interface datailData {
-  nickname: string
-  title: string
-  content: string
-  category: string
-  meetingTime: string
-  market: string
-  recruitmentNumber: number
-  applyNumber: number
-  uploadTime: string
-}
-
-// interface RouteState {
-//   state: {
-//     category: string
-//   }
-// }
-
-const DetailContent = ({ id }: Props) => {
-  // const { state } = useLocation() as RouteState
-  const navigate = useNavigate()
+const DetailContent = ({ data }: Props) => {
   const [isJoin, setIsJoin] = useState(false)
   const [isLike, setIsLike] = useState(false)
-  const [result, setResult] = useState<datailData>()
   const { isOpen, onClose, setIsOpen } = useModal()
 
   const likeToggle = () => {
@@ -46,22 +25,7 @@ const DetailContent = ({ id }: Props) => {
     onClose()
   }
 
-  useEffect(() => {
-    const getDetailContent = async () => {
-      try {
-        const { data, status } = await authInstance.get(`post/${id}`)
-        if (status === 200) {
-          setResult(data)
-        }
-      } catch (error: any) {
-        if (error.response.status === 404) {
-          navigate('/error')
-        }
-      }
-      return {}
-    }
-    getDetailContent()
-  }, [id, navigate])
+  if (!data) return null
 
   return (
     <div className={styles.detailContent}>
@@ -70,29 +34,29 @@ const DetailContent = ({ id }: Props) => {
       </div>
       <div className={styles.detailNickname}>
         <span className={styles.profile} />
-        <span className={styles.nickname}>{result?.nickname}</span>
+        <span className={styles.nickname}>{data.nickname}</span>
       </div>
       <div className={styles.detailTitle}>
-        <h3>{result?.title}</h3>
+        <h3>{data.title}</h3>
       </div>
       <div className={styles.detailText}>
-        <p>{result?.content}</p>
+        <p>{data.content}</p>
       </div>
       <div className={styles.detailList}>
         <ul>
           <li className={styles.location}>
             <LocationIcon />
-            {result?.market}
+            {data.market}
           </li>
           <li className={styles.time}>
             <TimeIcon />
-            {result?.meetingTime}
+            {data.meetingTime}
           </li>
           <li className={styles.people}>
             <PeopleIcon />
-            {result?.applyNumber} / {result?.recruitmentNumber}
+            {data.applyNumber} / {data.recruitmentNumber}
           </li>
-          <li>{result?.uploadTime}</li>
+          <li>{data.uploadTime}</li>
         </ul>
       </div>
       <div className={styles.detailBtn}>
