+++ src/pages/Detail/index.tsx
@@ -1,6 +1,8 @@
-import { useParams } from 'react-router-dom'
+import { useQuery } from 'react-query'
+import { useParams, useNavigate } from 'react-router-dom'
 
 import useModal from 'hooks/useModal'
+import { authInstance } from 'apis/client'
 import { MoreIcon } from 'assets/svgs'
 
 import { UnderModal } from 'components/Modal'
@@ -11,9 +13,16 @@ import Header from 'components/Header'
 
 import styles from './detail.module.scss'
 
+const getDetailAPI = (id: string | undefined) => authInstance.get(`post/${id}`)
+
 const Detail = () => {
-  const { isOpen, onClose, setIsOpen } = useModal()
   const { id } = useParams()
+  const { data } = useQuery(['getDetailAPI', id], () => getDetailAPI(id).then((res) => res.data))
+  const navigate = useNavigate()
+  const { isOpen, onClose, setIsOpen } = useModal()
+
+  if (!data) return null
+
   return (
     <div className={styles.detail}>
       <div className={styles.headerWrap}>
@@ -27,14 +36,19 @@ const Detail = () => {
         />
       </div>
       <div className='contentsInner'>
-        <DetailContent id={id} />
+        <DetailContent data={data} />
         <Comment />
       </div>
       <UnderModal
         show={isOpen}
         close={onClose}
         items={[
-          { name: '수정', callback: () => {} },
+          {
+            name: '수정',
+            callback: () => {
+              navigate('/edit', { state: { data } })
+            },
+          },
           { name: '삭제', callback: () => {} },
         ]}
       />
